# apps/backend/openapi.json
{
...
  "openapi": "3.0.3",
  "info": {
    "title": "Minecraft MCP Server API",
    "version": "0.1.0",
    "description": "Model Context Protocol (MCP) server for Minecraft command execution and resource management",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "contact": {
      "name": "Minecraft MCP Server",
      "url": "https://github.com/olddude/minecraft-mcp-server"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/tools/execute-command": {
      "post": {
        "summary": "Execute Minecraft command",
        "description": "Execute a single Minecraft server command",
        "tags": ["Tools"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["command"],
                "properties": {
                  "command": {
                    "type": "string",
                    "description": "Minecraft command to execute",
                    "example": "time set day"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "output": {
                      "type": "string",
                      "description": "Command output"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/chat/completions": {
      "post": {
        "summary": "Create chat completion (OpenAI compatible)",
        "description": "Create a chat completion that bridges to Minecraft MCP tools",
        "tags": ["OpenAI"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenAIChatRequest"
              },
              "examples": {
                "minecraft-command": {
                  "summary": "Execute Minecraft commands",
                  "value": {
                    "messages": [
                      {
                        "role": "user",
                        "content": "Execute /time set day and /weather clear"
                      }
                    ],
                    "model": "minecraft-mcp"
                  }
                },
                "minecraft-help": {
                  "summary": "Get Minecraft help",
                  "value": {
                    "messages": [
                      {
                        "role": "user",
                        "content": "How do I build a house in Minecraft?"
                      }
                    ],
                    "model": "minecraft-mcp"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "OpenAIChatRequest": {
        "type": "object",
        "required": ["messages"],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpenAIMessage"
            },
            "description": "Chat conversation messages"
          },
          "model": {
            "type": "string",
            "description": "Model to use for completion",
            "default": "minecraft-mcp"
          },
          "stream": {
            "type": "boolean",
            "description": "Whether to stream responses",
            "default": false
          },
          "temperature": {
            "type": "number",
            "minimum": 0,
            "maximum": 2,
            "description": "Sampling temperature"
          },
          "max_tokens": {
            "type": "integer",
            "minimum": 1,
            "description": "Maximum tokens to generate"
          }
        }
      },
      "OpenAIMessage": {
        "type": "object",
        "required": ["role", "content"],
        "properties": {
          "role": {
            "type": "string",
            "enum": ["system", "user", "assistant"],
            "description": "Message role"
          },
          "content": {
            "type": "string",
            "description": "Message content"
          },
          "tool_calls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpenAIToolCall"
            }
          }
        }
      },
      "OpenAIToolCall": {
        "type": "object",
        "required": ["id", "type", "function"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Tool call identifier"
          },
          "type": {
            "type": "string",
            "enum": ["function"]
          },
          "function": {
            "type": "object",
            "required": ["name", "arguments"],
            "properties": {
              "name": {
                "type": "string",
                "description": "Function name"
              },
              "arguments": {
                "type": "string",
                "description": "Function arguments as JSON string"
              }
            }
          }
        }
      },
      "OpenAIChatResponse": {
        "type": "object",
        "required": ["id", "object", "created", "model", "choices"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Completion identifier"
          },
          "object": {
            "type": "string",
            "enum": ["chat.completion"]
          },
          "created": {
            "type": "integer",
            "description": "Unix timestamp"
          },
          "model": {
            "type": "string",
            "description": "Model used"
          },
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpenAIChoice"
            }
          },
          "usage": {
            "$ref": "#/components/schemas/OpenAIUsage"
          }
        }
      },
      "OpenAIChoice": {
        "type": "object",
        "required": ["index", "message", "finish_reason"],
        "properties": {
          "index": {
            "type": "integer",
            "description": "Choice index"
          },
          "message": {
            "$ref": "#/components/schemas/OpenAIMessage"
          },
          "finish_reason": {
            "type": "string",
            "enum": ["stop", "length", "tool_calls"],
            "description": "Reason for completion finish"
          }
        }
      },
      "OpenAIUsage": {
        "type": "object",
        "required": ["prompt_tokens", "completion_tokens", "total_tokens"],
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "description": "Tokens in prompt"
          },
          "completion_tokens": {
            "type": "integer",
            "description": "Tokens in completion"
          },
          "total_tokens": {
            "type": "integer",
            "description": "Total tokens used"
          }
        }
      }
    },
    "securitySchemes": {
      "sessionId": {
        "type": "apiKey",
        "in": "header",
        "name": "Mcp-Session-Id",
        "description": "MCP session identifier for stateful connections"
      }
    }
  },
...
}